requirements.txt
pytest==8.2.2
pytest-dependency==0.6.0
pytest-order==1.2.1
selenium==4.22.0
requests==2.32.3
Pillow==10.4.0
numpy==2.0.0
configparser==7.0.0
beautifulsoup4==4.12.3
bs4==0.0.2
PyMySQL==1.1.1
lipsum==0.1.2
hvac==2.3.0
browsermob-proxy==0.8.0
gcloud==0.18.3
google-cloud-storage==2.17.0
pytest-xdist==3.6.1
pytest-rerunfailures==14.0
pytest-timeout==2.3.1
sshtunnel==0.4.0
pylint==3.2.5
pytest-html==4.1.1
google-api-python-client==2.136.0
google-auth==2.31.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
pytz==2024.1
redis==5.0.7
webdriver-manager==4.0.1